refactor(wallet): extract useTelegramMainButton hook
